+++ web/.gitignore
@@ -14,3 +14,5 @@ __screenshots__
 /cesium_ion_token.txt
 .idea/*
 .yarn/*
+e2e/utils/.auth
+e2e/testSnapshot
