+++ web/e2e/dashboard.spec.ts
@@ -1,3 +1,4 @@
+// not in use
 // import { expect, test } from "@reearth/e2e/utils";
 
 // test("dasboard can be logged in", async ({ page, reearth }) => {
