+++ web/e2e/pom/visualizer/editor/MapPage.ts
