+++ web/e2e/utils/login.ts
@@ -1,3 +1,4 @@
+//not in use
 import axios from "axios";
 
 import { config } from "./config";
