+++ web/e2e/utils/setup.ts
@@ -1,3 +1,4 @@
+//not in use
 import { getAccessToken, setAccessToken } from "./config";
 import { login } from "./login";
